+++ service-providers/network-requester/src/connection.rs
@@ -7,6 +7,7 @@ use proxy_helpers::connection_controller::ConnectionReceiver;
 use proxy_helpers::proxy_runner::ProxyRunner;
 use socks5_requests::{ConnectionId, Message as Socks5Message, RemoteAddress, Response};
 use std::io;
+use task::ShutdownListener;
 use tokio::net::TcpStream;
 
 /// A TCP connection between the Socks5 service provider, which makes
@@ -40,6 +41,7 @@ impl Connection {
         &mut self,
         mix_receiver: ConnectionReceiver,
         mix_sender: mpsc::UnboundedSender<(Socks5Message, Recipient)>,
+        shutdown: ShutdownListener,
     ) {
         let stream = self.conn.take().unwrap();
         let remote_source_address = "???".to_string(); // we don't know ip address of requester
@@ -52,6 +54,7 @@ impl Connection {
             mix_receiver,
             mix_sender,
             connection_id,
+            shutdown,
         )
         .run(move |conn_id, read_data, socket_closed| {
             (